Switch to JSON-encoding for the stored files.
The choices were to keep with the Go gob format, which isn't really
human editable, use json (which is in the Go standard library), or
find / implement another serialisation format. As bad as json is (c.f.
the rectify and compensate methods), it's less work for a tool that
I just need to work.